requirements.txt
pandas
htmlmin
yagmail
pyyaml
beautifulsoup4
google-cloud-storage
fsspec
gcsfs